AverageRating now uses DoubleWritable instead of IntWritable for more accurate ratings

+++ average-rating/src/main/java/com/average_rating/AverageRatingDriver.java
@@ -2,7 +2,7 @@ package main.java.com.average_rating;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -20,10 +20,10 @@ public class AverageRatingDriver {
 
         // Set Input and Output Parameters
         job.setMapOutputKeyClass(Text.class);
-        job.setMapOutputValueClass(IntWritable.class);
+        job.setMapOutputValueClass(DoubleWritable.class);
 
         job.setOutputKeyClass(Text.class);
-        job.setOutputValueClass(IntWritable.class);
+        job.setOutputValueClass(DoubleWritable.class);
 
         // Number of Reducers
         job.setNumReduceTasks(1);
